fix #311950: add .mpos/.spos-like exports whose units are PNG pixels
• name these formats .mpx and .spx for m̲easure/s̲egment p̲ositions
  in pixels (hence the px) ☻
• add support in both places where MuseScore exports files
• in mpx/spx rename sx and sy attributes to w and h to make it
  clearer what they actually are
• add extra information to top-level <score/> tag:
  – posformat: 1 (old format) or 2 (new pixel-based format)
  – kind: "segment" or "measure" respectively
  – width, height: by request of @dmitrio95 the size of a page

The page size units are scaling units, that is, "width" is the
maximum number an element’s x+sx or x+w could ever become (so
this scales with the element position scaling factor and helps
to position things relative to the whole (page) canvas).

Adding the extra XML attributes also for format 1 should not
break a reader: they’d just ignore unknown attributes.

Renaming the sx and sy attributes was also done to avoid
reading the new format in an old-format reader; new-format
readers should check the posformat attribute to be what
they expect (that is, "2").

Finally, pixel calculations are changed as follows:
• an offset (element@x and element@y) is always rounded down
• a width or height (sx/sy, w/h) is always rounded up

The idea here is that even a fractional pixel can be set,
especially with anti-aliasing in play.

The PNG export code is also changed to use qCeil() instead
of lrint() [round to nearest neighbour] for calculating
the PNG size, to match this and never lose a possible, even
partial, pixel.